add bw_{and/or/xor}_uf

+++ tests/cn/memcpy_2.error.c
@@ -1,6 +1,6 @@
 
 /*@
-predicate (u8) Owned_char_wrapper(pointer p) {
+predicate (integer) Owned_char_wrapper(pointer p) {
     take c = RW<char>(p);
     return c;
 }
@@ -18,35 +18,35 @@ triggering Fulminate's optimisations for conjunctions and checking they work in 
 */
 void
 naive_memcpy (char *dst, char *src, int n)
-/*@ requires take dstStart_half_1 = each (i32 j; 0i32 <= j && j < n && j % 2i32 == 0i32)
+/*@ requires take dstStart_half_1 = each (integer j; 0 <= j && j < n && j % 2 == 0)
                                   {RW(array_shift(dst, j))};
-             take dstStart_half_2 = each (i32 j; 0i32 <= j && j < n && j % 2i32 == 1i32)
+             take dstStart_half_2 = each (integer j; 0 <= j && j < n && j % 2 == 1)
                                   {Owned_char_wrapper_void(array_shift(dst, j))};
-             take srcStart = each (i32 j; 0i32 <= j && j < n)
+             take srcStart = each (integer j; 0 <= j && j < n)
                                   {Owned_char_wrapper(array_shift(src, j))};
-    ensures take dstEnd = each (i32 j; 0i32 <= j && j < n)
+    ensures take dstEnd = each (integer j; 0 <= j && j < n)
                                {RW(array_shift(dst, j))};
-            take srcEnd = each (i32 j; 0i32 <= j && j < n)
+            take srcEnd = each (integer j; 0 <= j && j < n)
                                {RW(array_shift(src, j))};
             srcEnd == srcStart;
-            each (i32 k; 0i32 <= k && k < n) {dstEnd[k] == srcStart[k]};
+            each (integer k; 0 <= k && k < n) {dstEnd[k] == srcStart[k]};
 @*/
 {
   int i;
   for (i = 0; i < n; i = i + 1)
-  /*@ inv take dstInv = each (i32 j; 0i32 <= j && j < n)
+  /*@ inv take dstInv = each (integer j; 0 <= j && j < n)
                              {RW(array_shift(dst, j))};
-          take srcInv = each (i32 j; 0i32 <= j && j < n)
+          take srcInv = each (integer j; 0 <= j && j < n)
                              {RW(array_shift(src, j))};
           // srcInv == srcStart;
-          each (i32 j; 0i32 <= j && j < i) {dstInv[j] == srcStart[j]};
-          0i32 <= i;
+          each (integer j; 0 <= j && j < i) {dstInv[j] == srcStart[j]};
+          0 <= i;
           {dst} unchanged;
           {src} unchanged;
           {n} unchanged; @*/
   {
-    /*@ focus RW<char>, (i32)i; @*/
-    /*@ instantiate good<char>, (i32)i; @*/
+    /*@ focus RW<char>, i; @*/
+    /*@ instantiate good<char>, i; @*/
     dst[i] = src[i];
   }
 }
